Improve implementation based on GDTF Spec 1.2  (#19)
This adds new implementation for some parts and removes other for stuff removed
from the spec. There should be no breaking changes for version 1.2 GDTF files.
Some of the stuff existing in GDTF 1.0/1.1 that has been removed from the
specification and made it to this library (for example visual macros) has been
removed but should not affect anyone.


* Add Animation system

* Add Gamuts definition

* Add DMX Profiles implementation

* Add missing stuff to Beam Geometry

* Add missing parts to WiringObjects

* Add SubPhysicalUnits parsing

* Add SubChannelSet parsing

* Add missing channel function attributes

* Fix protocols

* Split into multiple files

* Fix protocols

* Fix Revisions

* Update macro to current spec

* isort imports

* Add attribution

* Remove unused commits

* Update readme